Center drive failed; pipe jumping still iffy
Tried a strategy to drive through the center, but it's pretty sketch as
to whether it would finish in time...or even finish at all.

Made changes to the basket to where it should (theoretically) be able to
get the tribbles properly, but it still has issues getting to the
dumping location and dumping properly.